Squashed 'luarocks/' changes from 04bbe2e..cae940b
cae940b Merge branch '3.1.2'
640d6ca Release 3.1.2
92a6aaa Tests: add smoke test
923f373 cmd: fix handling of unconfigured paths in get_lua_version
9cb923d add publishrelease script
5057903 Add mergerelease script
cd1a762 Merge branch '3.1.1'
7386679 Release 3.1.1
37ac8cd Tests: add smoke test
89c499c Add rockspec
191c170 fix: handle errors in lua wrapper check
8cfe305 refactor: reorganise code in get_lua_version
ea1e427 fix: don't try to look in home_tree if not available
33b00b4 Merge branch '3.1.0'
0585250 Release 3.1.0
8a8622b Tests: add smoke test
c451dd0 Tests: cleanup signature before signing
d534a32 Fix build of binaries
18b9a94 Tests: add test for fs.wrap_script
bb88566 bin wrapper: revert from LUA_INIT back to -e
e70441f hardcode "windows" platform in Windows binary to help Wine
67660ac travis: use pip (not pip2) with hererocks
5641b64 repos: prevent fs.move from failing
0171e22 patch lockfile: do not rely on relative paths
0c5f8ed fix: handle all-negative supported_platforms rockspec field
6000c9a fix: correctly print current platforms
beb38fe make: use a lockfile for patching on first `luarocks make` run
0b6f834 spec/init_spec.lua: cleanup test
7807b77 init: robustness improvements
3707b02 fs: separate cache_timeout and cache_fail_timeout
208c552 search: add report and opt-out for checking other Lua versions
29513ca persist: save_default_lua_version
dc2ecc0 Improve runtime detection of Lua interpreter
06f81d9 cmd: Fix use of sysconfdir for default-lua-version.lua search
f644e99 init: rewrite project config when it's safe to do so
a21fcdc config: add modes for reading, writing and unsetting individual entries
9c03508 Add --global, improve relationship between project and --local
0b80c36 Simplify issue with superuser cache and avoid annoying message
dec452e manif: use temp dir if local cache can't be created
d2423f4 util: more informative message: "Lua 5.x interpreter not found"
f5eef8f fs.lua: cache failures as well
b3d3651 Cache manifest more aggressively
1ebc96a update to luacov 0.13.1 + cluacov 0.1.1
38c02b5 util: fix refactor mistake
b89ab59 Reorder Travis tests
12562d3 install, build, make: add --verify and --sign options
b46978f Refactor opts_table from build to reuse in install
18842a0 upload: add --sign option
655eacf Add --sign option to `luarocks pack`
4e97804 tar: do not attempt to create directory more than once
7e9b7e0 fs: versions of exists, is_file, is_dir for Unix and Windows that do not fork
8283b81 init: ensure --reset resets lua wrapper and config file
fbb3431 fun: add find and memoize functions
e388dee cmd: refactor detection code and improve detection of project dir
5cdaa0a Add rockspec
a05899a upload: accept non-200 codes
973655f appveyor: improve caching
ae114e2 Remove spurious arguments
e4315e0 core.cfg: produce quoted values in tool variables for Windows
6d7723a deps: improve performance of external dependency checks
b7ad112 cmd: improve detection of project dir
9064b9f refactor: move require_json from luarocks.upload.api to luarocks.util
83dceb3 util: when bootstrapped, return self path as wrapper
26b2225 init: support running from the root directory
1219cb3 Do not register system tree if installing to ~/.luarocks
3855e59 core.cfg: make detection of sysconfdir version independent
f54728b cmd: be more lenient when --lua-version can't find an interpreter
63c5635 cmd: if ./.luarocks/config-5.x.lua exists, assume Lua version 5.x by default
f724ec6 Add info about lua-version option in help
788f1fa fixup! core.sysdetect: Fallback to /proc to detect Linux
9d87daf core.sysdetect: Fallback to /proc to detect Linux
b096e35 Support installations without documentation
28f1f87 fix: make variable changes local
7a28056 Fix small misspellings (#957)
27c0625 core.cfg: fix sysconfdir detection trimming @ from debug.getinfo
9bb5c28 persist: handle keys that are keywords correctly.
914a2b7 fs.win32: fix setting permissions for all users (#943)
dc46a81 only report on local cache ownership if we can check it
aafc594 makedist: add 'binary' and 'sign' steps
8707522 Merge branch '3.0.4'

git-subtree-dir: luarocks
git-subtree-split: cae940b7e3a89ff1cabc5752c30a138e9e52729b

+++ src/luarocks/manif.lua
@@ -72,13 +72,18 @@ local function fetch_manifest_from(repo_url, filename)
    local cache_dir = dir.path(cfg.local_cache, name)
    local ok = fs.make_dir(cache_dir)
    if not ok then
-      return nil, "Failed creating temporary cache directory "..cache_dir
+      cfg.local_cache = fs.make_temp_dir("local_cache")
+      cache_dir = dir.path(cfg.local_cache, name)
+      ok = fs.make_dir(cache_dir)
+      if not ok then
+         return nil, "Failed creating temporary cache directory "..cache_dir
+      end
    end
-   local file, err, errcode = fetch.fetch_url(url, dir.path(cache_dir, filename), true)
+   local file, err, errcode, from_cache = fetch.fetch_url(url, dir.path(cache_dir, filename), true)
    if not file then
       return nil, "Failed fetching manifest for "..repo_url..(err and " - "..err or ""), errcode
    end
-   return file
+   return file, nil, nil, from_cache
 end
 
 --- Load a local or remote manifest describing a repository.
@@ -106,7 +111,7 @@ function manif.load_manifest(repo_url, lua_version)
    }
 
    local protocol, repodir = dir.split_url(repo_url)
-   local pathname
+   local pathname, from_cache
    if protocol == "file" then
       for _, filename in ipairs(filenames) do
          pathname = dir.path(repodir, filename)
@@ -117,7 +122,7 @@ function manif.load_manifest(repo_url, lua_version)
    else
       local err, errcode
       for _, filename in ipairs(filenames) do
-         pathname, err, errcode = fetch_manifest_from(repo_url, filename)
+         pathname, err, errcode, from_cache = fetch_manifest_from(repo_url, filename)
          if pathname then
             break
          end
@@ -131,13 +136,15 @@ function manif.load_manifest(repo_url, lua_version)
       local dirname = dir.dir_name(pathname)
       fs.change_dir(dirname)
       local nozip = pathname:match("(.*)%.zip$")
-      fs.delete(nozip)
-      local ok, err = fs.unzip(pathname)
-      fs.pop_dir()
-      if not ok then
-         fs.delete(pathname)
-         fs.delete(pathname..".timestamp")
-         return nil, "Failed extracting manifest file: " .. err
+      if not from_cache then
+         fs.delete(nozip)
+         local ok, err = fs.unzip(pathname)
+         fs.pop_dir()
+         if not ok then
+            fs.delete(pathname)
+            fs.delete(pathname..".timestamp")
+            return nil, "Failed extracting manifest file: " .. err
+         end
       end
       pathname = nozip
    end
